refactor(utils): move GetFormattedHexdump to a dedicated utils package for better code organization
